Moved role files into dedicated directory.
