+++ frontend/src/routes/UserManagement/Roles/Role/RoleDetail.tsx
@@ -1,7 +1,7 @@
 /* Copyright Contributors to the Open Cluster Management project */
 import { PageSection } from '@patternfly/react-core'
 //import { useTranslation } from '../../../lib/acm-i18next'
-import { useCurrentRole } from './RolesPage'
+import { useCurrentRole } from '../RolesPage'
 
 const RoleDetail = () => {
   //const { t } = useTranslation()
