+++ frontend/src/routes/UserManagement/Roles/Role/RolePage.tsx
@@ -1,7 +1,7 @@
 /* Copyright Contributors to the Open Cluster Management project */
 import { useParams, useLocation, Link, Outlet, useNavigate } from 'react-router-dom-v5-compat'
-import { useTranslation } from '../../../lib/acm-i18next'
-import { useRolesContext, useCurrentRole } from './RolesPage'
+import { useTranslation } from '../../../../lib/acm-i18next'
+import { useRolesContext, useCurrentRole } from '../RolesPage'
 import {
   AcmPage,
   AcmPageHeader,
@@ -9,12 +9,12 @@ import {
   AcmSecondaryNavItem,
   AcmLoadingPage,
   AcmButton,
-} from '../../../ui-components'
-import { NavigationPath } from '../../../NavigationPath'
+} from '../../../../ui-components'
+import { NavigationPath } from '../../../../NavigationPath'
 import { generatePath } from 'react-router-dom-v5-compat'
 import { Page, PageSection } from '@patternfly/react-core'
-import { ErrorPage } from '../../../components/ErrorPage'
-import { ResourceError, ResourceErrorCode } from '../../../resources/utils'
+import { ErrorPage } from '../../../../components/ErrorPage'
+import { ResourceError, ResourceErrorCode } from '../../../../resources/utils'
 
 const RolePage = () => {
   const { t } = useTranslation()
