+++ frontend/src/routes/UserManagement/Roles/Role/RolePermissions.test.tsx
@@ -3,7 +3,7 @@
 import { render, screen } from '@testing-library/react'
 import { MemoryRouter } from 'react-router-dom-v5-compat'
 import { RecoilRoot } from 'recoil'
-import { nockIgnoreRBAC } from '../../../lib/nock-util'
+import { nockIgnoreRBAC } from '../../../../lib/nock-util'
 import { RolePermissions } from './RolePermissions'
 
 function Component({ roleId = 'test-role' }: { roleId?: string }) {
