+++ frontend/src/routes/UserManagement/Roles/Role/RolePermissions.tsx
@@ -1,13 +1,13 @@
 /* Copyright Contributors to the Open Cluster Management project */
 import { PageSection } from '@patternfly/react-core'
 //import { useTranslation } from '../../../lib/acm-i18next'
-import { useCurrentRole } from './RolesPage'
+import { useCurrentRole } from '../RolesPage'
 
-const RoleYaml = () => {
+const RolePermissions = () => {
   //const { t } = useTranslation()
   const role = useCurrentRole()
 
-  return <PageSection>Role YAML page for Role: {role?.metadata.name}</PageSection>
+  return <PageSection>Role permissions page for Role: {role?.metadata.name}</PageSection>
 }
 
-export { RoleYaml }
+export { RolePermissions }
