+++ frontend/src/routes/UserManagement/Roles/Role/RoleYaml.test.tsx
@@ -3,7 +3,7 @@
 import { render, screen } from '@testing-library/react'
 import { MemoryRouter } from 'react-router-dom-v5-compat'
 import { RecoilRoot } from 'recoil'
-import { nockIgnoreRBAC } from '../../../lib/nock-util'
+import { nockIgnoreRBAC } from '../../../../lib/nock-util'
 import { RoleYaml } from './RoleYaml'
 
 function Component({ roleId = 'test-role' }: { roleId?: string }) {
